+++ app/src/actions/meActions.js
@@ -116,3 +116,9 @@ export const setAutoMuted = (flag) =>
 		type    : 'SET_AUTO_MUTED',
 		payload : { flag }
 	});
+
+export const setLocalRecordingState = (localRecordingState) =>
+	({
+		type    : 'SET_LOCAL_RECORDING_STATE',
+		payload : { localRecordingState }
+	});