+++ app/src/reducers/notifications.js
@@ -6,6 +6,8 @@ const notifications = (state = [], action) =>
 		{
 			const { notification } = action.payload;
 
+			notification.toBeClosed=false;
+
 			return [ ...state, notification ];
 		}
 
@@ -19,6 +21,21 @@ const notifications = (state = [], action) =>
 		case 'REMOVE_ALL_NOTIFICATIONS':
 		{
 			return [];
+		}
+
+		case 'CLOSE_NOTIFICATION':
+		{
+			const { notificationId } = action.payload;
+
+			return (state.map((e) =>
+			{
+				if (e.id === notificationId)
+				{
+					e.toBeClosed=true;
+				}
+
+				return e;
+			}));
 		}
 
 		default: