+++ app/src/store.js
@@ -88,7 +88,7 @@ if (process.env.REACT_APP_DEBUG === '*' || process.env.NODE_ENV !== 'production'
 			duration  : true,
 			collapsed : true,
 			timestamp : false,
-			level     : 'log',
+			level     : 'info',
 			logErrors : true
 		});