+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 3.5.0
+
+### Added
+
+* Local Recording
+  * IndexedDB and fallback to memory (array of blobs)
+  * CLIENT.recoverRecording(`DB_NAME`)
+  * Add new permission LOCAL_RECORD_ROOM
 ## 3.5.0 or Next Version
 
 ### Added